+++ components/tooltip/tests/tooltip.spec.tsx
@@ -39,7 +39,7 @@ describe('Tooltip', () => {
     wrapper = createTooltip()
 
     await wrapper.find('.trigger').trigger('mouseenter')
-    vi.runAllTimers()
+    vi.runOnlyPendingTimers()
     await nextTick()
     expect(wrapper.find('.vxp-tooltip__popper').exists()).toBe(true)
     expect(wrapper.find('.vxp-tooltip__popper').classes()).toContain('vxp-tooltip-vars')
@@ -51,7 +51,7 @@ describe('Tooltip', () => {
     wrapper = createTooltip({ trigger: 'click' })
 
     await wrapper.find('.trigger').trigger('mouseenter')
-    vi.runAllTimers()
+    vi.runOnlyPendingTimers()
     await nextTick()
     expect(wrapper.find('.vxp-tooltip__popper').exists()).toBe(false)
 
@@ -65,7 +65,7 @@ describe('Tooltip', () => {
     wrapper = createTooltip({ trigger: 'focus' })
 
     await wrapper.find('.trigger').trigger('mouseenter')
-    vi.runAllTimers()
+    vi.runOnlyPendingTimers()
     await nextTick()
     expect(wrapper.find('.vxp-tooltip__popper').exists()).toBe(false)
 
@@ -79,24 +79,24 @@ describe('Tooltip', () => {
     wrapper = createTooltip({ trigger: 'hover-focus' })
 
     await wrapper.find('.trigger').trigger('mouseenter')
-    vi.runAllTimers()
+    vi.runOnlyPendingTimers()
     await nextTick()
     expect(wrapper.find('.vxp-tooltip__popper').exists()).toBe(true)
     expect(wrapper.find('.tip').exists()).toBe(true)
     expect(wrapper.find('.tip').text()).toEqual(TEXT)
 
     await wrapper.find('.trigger').trigger('focus')
-    vi.runAllTimers()
+    vi.runOnlyPendingTimers()
     await nextTick()
     expect(wrapper.find('.vxp-tooltip__popper').exists()).toBe(true)
 
     await wrapper.find('.trigger').trigger('mouseleave')
-    vi.runAllTimers()
+    vi.runOnlyPendingTimers()
     await nextTick()
     expect(wrapper.find('.vxp-tooltip__popper').exists()).toBe(true)
 
     await wrapper.find('.trigger').trigger('blur')
-    vi.runAllTimers()
+    vi.runOnlyPendingTimers()
     await nextTick()
     expect(wrapper.find('.vxp-tooltip__popper').exists()).toBe(false)
   })
@@ -105,7 +105,7 @@ describe('Tooltip', () => {
     wrapper = createTooltip({ trigger: 'custom' })
 
     await wrapper.find('.trigger').trigger('mouseenter')
-    vi.runAllTimers()
+    vi.runOnlyPendingTimers()
     await nextTick()
     expect(wrapper.find('.vxp-tooltip__popper').exists()).toBe(false)
 
@@ -121,7 +121,7 @@ describe('Tooltip', () => {
     expect(wrapper.find('.trigger').exists()).toBe(true)
 
     await wrapper.find('.trigger').trigger('mouseenter')
-    vi.runAllTimers()
+    vi.runOnlyPendingTimers()
     await nextTick()
     await wrapper.setProps({ disabled: true })
     expect(wrapper.find('.vxp-tooltip__popper').exists()).toBe(false)
@@ -131,7 +131,10 @@ describe('Tooltip', () => {
     wrapper = createTooltip({ transfer: true })
 
     await wrapper.find('.trigger').trigger('mouseenter')
-    vi.runAllTimers()
+    vi.runOnlyPendingTimers()
+    await nextTick()
+    await nextTick()
+    vi.runOnlyPendingTimers()
     await nextTick()
     await nextTick()
     vi.runAllTimers()