+++ components/tree/tree-node.vue
@@ -72,7 +72,7 @@ useModifier({
 })
 
 const loaded = ref(!treeState.boundAsyncLoad || props.node.loaded)
-const loadFail = ref(!treeState.boundAsyncLoad || props.node.loadFail)
+const loadFail = ref(treeState.boundAsyncLoad && props.node.loadFail)
 const dragging = ref(false)
 const isDragOver = ref(false)
 const focused = ref(false)
@@ -173,7 +173,7 @@ onBeforeUnmount(() => {
 // }
 
 function setValue<T = unknown>(key: keyof TreeNodeProps, value: T) {
-  (props.node as any)[key] = value
+  ;(props.node as any)[key] = value
 }
 
 function handleClick() {
@@ -196,7 +196,15 @@ function handleToggleCheck(able = !props.node.checked) {
 }
 
 async function toggleExpanded(able = !props.node.expanded) {
-  if (props.node.loading || isDisabled.value || props.node.expandDisabled || isLeaf.value) return
+  if (
+    treeState.expanding ||
+    props.node.loading ||
+    isDisabled.value ||
+    props.node.expandDisabled ||
+    isLeaf.value
+  ) {
+    return
+  }
 
   if (able && treeState.boundAsyncLoad && !loaded.value) {
     setValue('loading', true)
@@ -332,23 +340,25 @@ function handleDragEnd(event: DragEvent) {
       :node="node"
       :depth="node.depth"
       :focused="focused"
-      :line-count="node.depth - node.inLastCount"
+      :line-count="0"
+      :line-indexes="node.lineIndexes"
       :toggle-check="handleToggleCheck"
       :toggle-expand="toggleExpanded"
       :toggle-select="handleToggleSelect"
     >
       <template v-if="hasLinkLine">
         <div
-          v-for="n in node.depth - node.inLastCount"
-          :key="n"
+          v-for="(lineIndex, index) in node.lineIndexes"
+          :key="index"
           :class="[
             nh.be('link-line'),
             nh.bem('link-line', 'vertical'),
-            n === 1 && nh.bem('link-line', 'first')
+            !index && nh.bem('link-line', 'first')
           ]"
-          :style="{ [nh.cv('link-line-index')]: n - 1 }"
+          :style="{ [nh.cv('link-line-index')]: lineIndex }"
           aria-hidden="true"
         ></div>
+
         <div
           :class="[nh.be('link-line'), nh.bem('link-line', 'horizontal')]"
           aria-hidden="true"